Add desktop file from Google Chrome, without renaming it
The original idea was to rename this file into eos-google-chrome.desktop
but that won't play nicely when setting the browser as the default one,
since Google Chrome internally checks for the name of the desktop file
and won't be able to realize that it's actually set as the default browser
if not matching the name.

This should not be a problem in our case since we won't be exposing the
original desktop file from Chrome via flatpak in any case, and the only
desktop file that will be installed with this name, in a discoverable
location, will be this one. So, it should be fine.

https://phabricator.endlessm.com/T14100