Move most of the logic to launch Google Chrome out of the helper
The reasoning behind this is that, otherwise, we'd need to ship OS
updates each time we needed to change/update how the launcher works
for a specific version of Chrome, as well as when we migrate into
the upstream solution for external flatpak apps.

Thus, from now on we will keep here the simplest logic to just look
for the external flatpak app with Google Chrome, where we'd expect
to find the actual launcher script that knows how to run the app
from outside flatpak's sandbox, setting environment variables and
anything else it might need, as it sees fit.

https://phabricator.endlessm.com/T14244