Use the system helper to touch the system-wide stamp file
Unfortunately, gnome-session's AutostartCondition uses ~/.config
as the base directory to find the files specified in that key,
so this means that we can't prevent the execution of the installer
script with the system-wide one (see GNOME bug 777994).

That should not be a big problem for now, though, as the installer
script will early return anyway if the system-wide file is found,
but it would be nice to re-add the AutostartCondition key to the
desktop file if it gnome-session ever supports absolute paths.

https://phabricator.endlessm.com/T15308