Refactor how we install files in the system, and use a wrapper shell script
The way things are currently laid out is a bit strange, with the main python
script used to launch Google Chrome in $(bindir) and other python scripts and
files (e.g. the installer and the config.py files) placed inside $(datadir),
which besides being unintuitive is also problematic (e.g. include files).

This patch re-organizes where the python scripts are installed (now all of
them will be under $(datadir)), providing a wrapper shell script to be
installed in $(bindir), as it's common practice, and adapt all the rest
of the files to acknowledge this new paths and file names.

https://phabricator.endlessm.com/T15308