Renamed source debian package to eos-browser-tools
This is done in preparation for the split that we are going to go for,
that will allow us to ship the eos-google-chrome-helper package separately
from other browser-related tools, that will be shipped as a separate package.

Note that this commit also includes the necessary changes to the autotools
infrastructure so that the files for the Chrome helper are still installed
in where they used to be, with generated files pointing to the same paths.

https://phabricator.endlessm.com/T15917