Added a new eos-browser-tools package with a handler for webapp:// URIs
This is a new package to ship this helper script outside of the chromium
package, but without getting it together with eos-google-chrome-helper,
which serves a completely separate purpose.

Also, split the files of the now secondary eos-google-chrome-helper package
out to its own source code directory, to make things less confusing.

https://phabricator.endlessm.com/T15917