Update .desktop file
I compared the file shipped in Google's .deb:

    $(flatpak info -l com.google.Chrome)/files/extra/usr/share/applications/google-chrome.desktop

with the one in this repository, and applied the applicable changes.
This should make it possible to choose New Window and New Private Window
by right-clicking the Chrome desktop icon. It also adds a number of
MIME type handlers, and (in particular) removes the claim to handle the
ftp:// URI scheme, which was removed from Chrome 91+.

https://phabricator.endlessm.com/T32562