chrome-helper: Drop support for auto-installing Chrome
Drop the `--initial-setup` argument from
`eos-google-chrome-installer.py`, and the auto-start desktop file which
launched that on first boot.

All Endless OS images now default to Chromium (T32407), so
auto-installing Chrome is no longer necessary. Until we can remove the
Chrome launcher from the icon grid by default (T32409), however, the
launcher still needs to be hooked up to a way to install Chrome, which
continues to be `eos-google-chrome-installer.py` (without the
`--initial-setup` argument).

The config file and stamp files for auto-installation are no longer
needed. The legacy stamp file for `eos-google-chrome-installer.py`, in
the user’s home directory, is unlinked directly. The system-wide stamp
file will be removed by a tmpfiles snippet in a following commit.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T32410