data: Hide google-chrome.desktop if Endless flatpak is not installed
Hide this desktop file if the Endless-provided Chrome flatpak is not installed.

This removes the feature (available in EOS <4.0) which always shows the Chrome
icon, and installs Chrome on demand if it’s clicked. Now, the icon will only
be shown if the Endless Chrome flatpak is installed. If the upstream (flathub)
flatpak is installed, this icon will not be shown — that flatpak provides its
own desktop file.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T32409